Separated Core code into their own dedicated Management classes
