+++ BabyStepsMultiplayerClient/Patches/Patch_PlayerMovement.cs
@@ -11,12 +11,12 @@ namespace BabyStepsMultiplayerClient.Patches
         private static bool DropHandItem_Prefix(PlayerMovement __instance,
             int __0) // hat
         {
-            if (Core.thisInstance.client == null)
+            if (Core.networkManager.client == null)
                 return true;
 
-            Grabable heldItem = Core.basePlayerMovement.handItems[__0];
+            Grabable heldItem = Core.localPlayer.basePlayerMovement.handItems[__0];
             if (heldItem != null)
-                Core.thisInstance.SendDropGrabable(__0);
+                Core.networkManager.SendDropGrabable(__0);
 
             // Run Original
             return true;
@@ -26,10 +26,10 @@ namespace BabyStepsMultiplayerClient.Patches
         [HarmonyPatch(typeof(PlayerMovement), nameof(PlayerMovement.KnockOffHat))]
         private static bool KnockOffHat_Prefix(PlayerMovement __instance)
         {
-            if (Core.thisInstance.client == null)
+            if (Core.networkManager.client == null)
                 return true;
 
-            Core.thisInstance.SendDoffHat();
+            Core.networkManager.SendDoffHat();
 
             // Run Original
             return true;
@@ -40,17 +40,17 @@ namespace BabyStepsMultiplayerClient.Patches
         private static void WearHat_Postfix(PlayerMovement __instance, 
             Hat __0) // hat
         {
-            if (Core.thisInstance.client == null)
+            if (Core.networkManager.client == null)
                 return;
 
-            Core.thisInstance.SendDonHat(__0);
+            Core.networkManager.SendDonHat(__0);
 
             for (int i = 0; i < __instance.handItems.Length; i++)
             {
                 Grabable item = __instance.handItems[i];
                 if ((item != null)
                     && item.name.Contains(__0.name))
-                    Core.thisInstance.SendDropGrabable(i);
+                    Core.networkManager.SendDropGrabable(i);
             }
         }
     }
